+++ module/imagemodule/internal/model/repository.go
@@ -11,5 +11,5 @@ type ImageRepository interface {
 	ListImages(ctx context.Context, db *gorm.DB) ([]imagemodel.Image, error)
 	Update(ctx context.Context, tx *gorm.DB, entity *ImageEntity) error
 	Create(ctx context.Context, tx *gorm.DB, entity *ImageEntity) error
-	FindForUpdate(ctx context.Context, tx *gorm.DB, id int) (*ImageEntity, error)
+	FindForUpdate(ctx context.Context, tx *gorm.DB, id int64) (*ImageEntity, error)
 }
